renamed file + deprecate sync health responder
