+++ verifiers/serve/client/zmq_env_client.py
@@ -10,9 +10,8 @@ import zmq
 import zmq.asyncio
 
 from verifiers.utils.logging_utils import print_time
-from verifiers.utils.worker_utils import msgpack_encoder
+from verifiers.utils.serve_utils import derive_health_address, msgpack_encoder
 from verifiers.serve.client.env_client import EnvClient
-from verifiers.utils.worker_utils import derive_health_address
 from verifiers.serve.types import (
     BaseRequest,
     BaseResponseT,
