+++ verifiers/utils/serve_utils.py
@@ -47,36 +47,6 @@ def derive_health_address(address: str) -> str:
     return f"{prefix}:{int(port_str) + 1}"
 
 
-def run_health_responder(address: str, stop_event) -> None:
-    """Synchronous health check responder in a dedicated process.
-
-    Completely isolated from the main server process's GIL, so health
-    pings always receive a prompt response regardless of workload.
-    """
-    import msgpack
-    import zmq
-
-    ctx = zmq.Context()
-    sock = ctx.socket(zmq.REP)
-    sock.setsockopt(zmq.LINGER, 0)
-    sock.setsockopt(zmq.RCVTIMEO, 1000)
-    sock.bind(address)
-
-    resp = msgpack.packb({"success": True, "error": None}, use_bin_type=True)
-
-    while not stop_event.is_set():
-        try:
-            sock.recv()
-            sock.send(resp)
-        except zmq.Again:
-            continue
-        except zmq.ZMQError:
-            break
-
-    sock.close()
-    ctx.term()
-
-
 def get_free_port() -> int:
     """Get a free port on the system."""
     with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
